+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerInvalid.jsx
@@ -27,6 +27,7 @@ const DateTimePickerExampleInvalid = () => {
           dateInputAttributes={{ id: 'default-invalid' }}
           onChange={handleDateTimeChange}
           isInvalid={isInvalid}
+          errorId="default-invalid-error"
         />
       </Field>
     </div>