+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerTimeZone.jsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import DateTimePicker from 'terra-date-time-picker/lib/DateTimePicker';
 import moment from 'moment-timezone';
 import Field from 'terra-form-field';
+import Fieldset from 'terra-form-fieldset';
 import DateTimeUtils from 'terra-date-time-picker/lib/DateTimeUtils';
 
 function DateTimePickerTimeZone() {
@@ -21,20 +22,20 @@ function DateTimePickerTimeZone() {
         Selected ISO Date Time:
         {dateTime}
       </p>
-      <p>
-        Initial Timezone:
-        {timeZone}
-      </p>
-      <Field label="Enter Date/Time" htmlFor="timeZone">
-        <DateTimePicker
-          ariaLabel="Enter Date/Time"
-          name="date-time-picker-example"
-          dateInputAttributes={{ id: 'timezone' }}
-          onChange={handleDateTimeChange}
-          initialTimeZone={timeZone}
-          value={dateTime}
-        />
-      </Field>
+      <Fieldset
+        legend={`Initial Timezone: ${timeZone}`}
+      >
+        <Field label="Enter Date/Time" htmlFor="timeZone">
+          <DateTimePicker
+            ariaLabel="Enter Date/Time"
+            name="date-time-picker-example"
+            dateInputAttributes={{ id: 'timezone' }}
+            onChange={handleDateTimeChange}
+            initialTimeZone={timeZone}
+            value={dateTime}
+          />
+        </Field>
+      </Fieldset>
     </>
   );
 }