+++ packages/terra-navigation-side-menu/src/_MenuItem.jsx
@@ -38,52 +38,63 @@ const propTypes = {
    * Text display for the menu item.
    * */
   text: PropTypes.string,
+  /**
+   * @private
+   * Menu container Ref for menu items
+   * */
+  getMenuContainerRef: PropTypes.func,
 };
 
 class MenuItem extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { active: false, focused: false };
+    this.state = { active: false };
     this.handleKeyDown = this.handleKeyDown.bind(this);
-    this.handleKeyUp = this.handleKeyUp.bind(this);
-    this.handleOnBlur = this.handleOnBlur.bind(this);
     this.textRender = this.textRender.bind(this);
-  }
-
-  handleOnBlur() {
-    this.setState({ focused: false });
+    this.handleMenuItemRef = this.handleMenuItemRef.bind(this);
   }
 
   handleKeyDown(event) {
-    // Add active state to FF browsers
-    if (event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
-      this.setState({ active: true });
+    const MenuContainerRef = this.props.getMenuContainerRef();
+    const listMenuItems = MenuContainerRef && MenuContainerRef.querySelectorAll('[data-menu-item]');
+    const currentIndex = Array.from(listMenuItems).indexOf(event.target);
+    const lastIndex = listMenuItems.length - 1;
+
+    if (event.nativeEvent.keyCode === KeyCode.KEY_DOWN) {
+      const nextIndex = currentIndex < lastIndex ? currentIndex + 1 : 0;
+      if (listMenuItems && listMenuItems[nextIndex]) {
+        listMenuItems[nextIndex].focus();
+      }
+      if (this.props.onKeyDown) {
+        this.props.onKeyDown(event);
+      }
+      event.preventDefault();
     }
 
-    // Add focus styles for keyboard navigation
     if (event.nativeEvent.keyCode === KeyCode.KEY_SPACE || event.nativeEvent.keyCode === KeyCode.KEY_RETURN) {
-      this.setState({ focused: true });
+      // Add active state to FF browsers
+      this.setState({ active: true });
+      this.props.onKeyDown(event);
     }
 
-    if (this.props.onKeyDown) {
-      this.props.onKeyDown(event);
+    if (event.nativeEvent.keyCode === KeyCode.KEY_UP) {
+      // Remove active state from FF broswers
+      if (event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
+        this.setState({ active: false });
+      }
+      const previousIndex = currentIndex > 0 ? currentIndex - 1 : lastIndex;
+      if (listMenuItems && listMenuItems[previousIndex]) {
+        listMenuItems[previousIndex].focus();
+      }
+      if (this.props.onKeyUp) {
+        this.props.onKeyUp(event);
+      }
+      event.preventDefault();
     }
   }
 
-  handleKeyUp(event) {
-    // Remove active state from FF broswers
-    if (event.nativeEvent.keyCode === KeyCode.KEY_SPACE) {
-      this.setState({ active: false });
-    }
-
-    // Apply focus styles for keyboard navigation
-    if (event.nativeEvent.keyCode === KeyCode.KEY_TAB) {
-      this.setState({ focused: true });
-    }
-
-    if (this.props.onKeyUp) {
-      this.props.onKeyUp(event);
-    }
+  handleMenuItemRef(node) {
+    this.contentNode = node;
   }
 
   textRender() {
@@ -116,7 +127,6 @@ class MenuItem extends React.Component {
       'menu-item',
       { 'is-selected': isSelected },
       { 'is-active': this.state.active },
-      { 'is-focused': this.state.focused },
       theme.className,
     ),
     customProps.className);
@@ -128,12 +138,11 @@ class MenuItem extends React.Component {
       >
         <div
           role="menuitem"
+          ref={this.handleMenuItemRef}
           {...customProps}
           tabIndex="0"
           className={itemClassNames}
           onKeyDown={this.handleKeyDown}
-          onKeyUp={this.handleKeyUp}
-          onBlur={this.handleOnBlur}
           aria-haspopup={hasChevron}
         >
           <div className={cx('title')}>