+++ packages/terra-navigation-side-menu/tests/wdio/navigation-side-menu-spec.js
@@ -42,4 +42,17 @@ Terra.describeViewports('Navigation Side Menu', ['medium'], () => {
       Terra.validates.element('isRootMenu navigation side menu display', { selector: '#test-menu' });
     });
   });
+
+  describe('Keyboard Navigation Using Arrow Keys', () => {
+    before(() => browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/navigation-side-menu-default'));
+
+    it('Navigates to first menu item', () => {
+      browser.keys(['Tab', 'Tab']);
+      Terra.validates.element('first menu item focused', { selector: '#root' });
+    });
+    it('should focus on second menu item', () => {
+      browser.keys(['ArrowDown']);
+      Terra.validates.element('second menu item focused', { selector: '#root' });
+    });
+  });
 });