+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.103.0 - (February 22, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 3.102.0 - (February 16, 2024)
 
 * Changed