+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+* Added
+  * Column actions example added to `terra-worklist-data-grid`.
+
+## 1.67.0 - (February 22, 2024)
+
+* Changed
+  * Updated the `terra-date-time-picker` timezone example.
+  * Updated selection examples and tests for `terra-folder-tree`.
+
 ## 1.66.0 - (February 20, 2024)
 
 * Changed
@@ -10,10 +19,13 @@
 ## 1.65.0 - (February 16, 2024)
 
 * Added
+  * Added ariaDescribedBy prop in `terra-date-time-picker` example for invalid error message.
+
+* Added
   * Added shouldTrapFocus to be set to 'true'.
-  * Column actions example added to `terra-worklist-data-grid`.
 
 * Changed
+  * Updated `terra-navigation-side-menu` example to use more meaningful labels.
   * Updated the `terra-date-time-picker` example for field label.
 
 ## 1.64.0 - (February 7, 2024)